+++ assets/system/game_modules/prim.mjs
@@ -55,14 +55,12 @@ class Prim
 		var v1 = 1.0 - sy / texture.height;
 		var u2 = (sx + width) / texture.width;
 		var v2 = 1.0 - (sy + height) / texture.height;
-		var shape = new Shape(ShapeType.TriStrip, texture,
-			new VertexList([
-				{ x: x1, y: y1, u: u1, v: v1, color: mask },
-				{ x: x2, y: y1, u: u2, v: v1, color: mask },
-				{ x: x1, y: y2, u: u1, v: v2, color: mask },
-				{ x: x2, y: y2, u: u2, v: v2, color: mask },
-			]));
-		shape.draw(surface);
+		Shape.drawImmediate(surface, ShapeType.TriStrip, texture, [
+			{ x: x1, y: y1, u: u1, v: v1, color: mask },
+			{ x: x2, y: y1, u: u2, v: v1, color: mask },
+			{ x: x1, y: y2, u: u1, v: v2, color: mask },
+			{ x: x2, y: y2, u: u2, v: v2, color: mask },
+		]);
 	}
 
 	static drawCircle(surface, x, y, radius, color)
@@ -87,9 +85,7 @@ class Prim
 				color: color
 			});
 		}
-		var vList = new VertexList(vertices);
-		var shape = new Shape(ShapeType.LineLoop, vList)
-		shape.draw(surface);
+		Shape.drawImmediate(surface, ShapeType.LineLoop, vertices);
 	}
 
 	static drawSolidCircle(surface, x, y, radius, color, color2)
@@ -122,9 +118,7 @@ class Prim
 			color: color2
 		};
 
-		var vList = new VertexList(vertices);
-		var shape = new Shape(ShapeType.Fan, vList);
-		shape.draw(surface);
+		Shape.drawImmediate(surface, ShapeType.Fan, vertices);
 	}
 
 	static drawSolidRectangle(surface, x, y, width, height, color_ul, color_ur, color_lr, color_ll)
@@ -133,14 +127,12 @@ class Prim
 		color_lr = color_lr || color_ul;
 		color_ll = color_ll || color_ul;
 
-		var shape = new Shape(ShapeType.TriStrip,
-			new VertexList([
-				{ x: x, y: y, color: color_ul },
-				{ x: x + width, y: y, color: color_ur },
-				{ x: x, y: y + height, color: color_ll },
-				{ x: x + width, y: y + height, color: color_lr },
-			]));
-		shape.draw(surface);
+		Shape.drawImmediate(surface, ShapeType.TriStrip, [
+			{ x: x, y: y, color: color_ul },
+			{ x: x + width, y: y, color: color_ur },
+			{ x: x, y: y + height, color: color_ll },
+			{ x: x + width, y: y + height, color: color_lr },
+		]);
 	}
 
 	static drawSolidTriangle(surface, x1, y1, x2, y2, x3, y3, color1, color2, color3)
@@ -148,19 +140,17 @@ class Prim
 		color2 = color2 || color1;
 		color3 = color3 || color1;
 
-		var shape = new Shape(ShapeType.Triangles,
-			new VertexList([
-				{ x: x1, y: y1, color: color1 },
-				{ x: x2, y: y2, color: color2 },
-				{ x: x3, y: y3, color: color3 },
-			]));
+		Shape.drawImmediate(surface, ShapeType.Triangles, [
+			{ x: x1, y: y1, color: color1 },
+			{ x: x2, y: y2, color: color2 },
+			{ x: x3, y: y3, color: color3 },
+		]);
 	}
 
 	static drawLine(surface, x1, y1, x2, y2, thickness, color1, color2)
 	{
 		color2 = color2 || color1;
 
-		var shape;
 		var xSize = x2 - x1;
 		var ySize = y2 - y1;
 		var length = Math.sqrt(xSize*xSize + ySize*ySize);
@@ -168,14 +158,12 @@ class Prim
 			return;
 		var tx = 0.5 * thickness * (y2 - y1) / length;
 		var ty = 0.5 * thickness * -(x2 - x1) / length;
-		shape = new Shape(ShapeType.Fan,
-			new VertexList([
-				{ x: x1 + tx, y: y1 + ty, color: color1 },
-				{ x: x1 - tx, y: y1 - ty, color: color1 },
-				{ x: x2 - tx, y: y2 - ty, color: color2 },
-				{ x: x2 + tx, y: y2 + ty, color: color2 },
-			]));
-		shape.draw(surface);
+		Shape.drawImmediate(surface, ShapeType.Fan, [
+			{ x: x1 + tx, y: y1 + ty, color: color1 },
+			{ x: x1 - tx, y: y1 - ty, color: color1 },
+			{ x: x2 - tx, y: y2 - ty, color: color2 },
+			{ x: x2 + tx, y: y2 + ty, color: color2 },
+		]);
 	}
 
 	static drawPoint(surface, x, y, color)
@@ -190,20 +178,18 @@ class Prim
 		var y1 = y + t;
 		var x2 = x1 + width - thickness;
 		var y2 = y1 + height - thickness;
-		var shape = new Shape(ShapeType.TriStrip,
-			new VertexList([
-				{ x: x1 - t, y: y1 - t, color: color },
-				{ x: x1 + t, y: y1 + t, color: color },
-				{ x: x2 + t, y: y1 - t, color: color },
-				{ x: x2 - t, y: y1 + t, color: color },
-				{ x: x2 + t, y: y2 + t, color: color },
-				{ x: x2 - t, y: y2 - t, color: color },
-				{ x: x1 - t, y: y2 + t, color: color },
-				{ x: x1 + t, y: y2 - t, color: color },
-				{ x: x1 - t, y: y1 - t, color: color },
-				{ x: x1 + t, y: y1 + t, color: color },
-			]));
-		shape.draw(surface);
+		Shape.drawImmediate(surface, ShapeType.TriStrip, [
+			{ x: x1 - t, y: y1 - t, color: color },
+			{ x: x1 + t, y: y1 + t, color: color },
+			{ x: x2 + t, y: y1 - t, color: color },
+			{ x: x2 - t, y: y1 + t, color: color },
+			{ x: x2 + t, y: y2 + t, color: color },
+			{ x: x2 - t, y: y2 - t, color: color },
+			{ x: x1 - t, y: y2 + t, color: color },
+			{ x: x1 + t, y: y2 - t, color: color },
+			{ x: x1 - t, y: y1 - t, color: color },
+			{ x: x1 + t, y: y1 + t, color: color },
+		]);
 	}
 
 	static fill(surface, color)